feat: request mp4 data when playing videos
this is because `dash` NEEDS mp4 media (instead of webm).

Because when playing videos, 2 URL sources are used (i.e. the audio+video), we consolidate those into a dash manifest, and that dash manifest requires that the URLs be mp4. In other cases, we select the opus URLs as they are better in any other imaginable sense.

For dash-webm, I hope gets some traction soon https://gstreamer.freedesktop.org/projects/dashwebm.html